+++ backend/ignition/modules/RaffleUpgradeable.ts
@@ -1,0 +1,7 @@
+import { buildModule } from '@nomicfoundation/hardhat-ignition/modules';
+
+export default buildModule('RaffleUpgradeable', (m) => {
+  const raffle = m.contract('Raffle', []);
+
+  return { raffle };
+});